Refactor layout structure and components
