+++ src/front/modules/layout/Layout.jsx
@@ -1,7 +1,7 @@
 import { Outlet }  from "react-router-dom"
-import ScrollToTop from "../components/ScrollToTop"
-import { Navbar }  from "../components/Navbar"
-import { Footer }  from "../components/Footer"
+import ScrollToTop from "./layout-subcomponents/ScrollToTop"
+import { Navbar }  from "./layout-subcomponents/Navbar"
+import { Footer }  from "./layout-subcomponents/Footer"
 
 // Base component that maintains the navbar and footer throughout the page and the scroll to top functionality.
 export const Layout = () => {
